Ensure backdrop filter cc::Layers are marked as drawing content
ContentLayerClientImpl::UpdateCcPictureLayer needs to set
cc::Layer::DrawsContent to true for content under a backdrop filter.
The previous code would do this based on the PendingLayer's effect node,
but this is incorrect if a single object has both a backdrop filter
effect node and a filter effect node. This approach failed to re-run
UpdateCcPictureLayer when HasBackdropEffect changed (see new WPT test
repaint-added-backdrop-filter.html).

This patch forces draws-content of all PendingLayers under an effect
node needing draws-content.

This patch is based on pdr@'s crrev.com/c/3367277. That patch uses
ScopedForceDrawsContent during painting and forces draws content flag
of PaintChunks. The problem is that we need to repaint the subtree of
an effect node on change of the node's MayDrawContent() status.

Bug: 1273849
Change-Id: Ie49d40c37767265e45278cf89ddf032895b9e832