Correct NGPhysicalFragment::IsTextControlPlaceholder().
We need to check that the fragment is a CSS box (and not e.g. a
fragmentainer), or NGContainerFragmentBuilder::AddChildInternal() would
insert fragmentainers in the wrong order (what should have been the
first fragmentainer would get stuck as the last one).

Also fix NGPhysicalFragment::IsTextControlContainer(), although with the
current usage it wasn't causing any problems.

Bug: 1295976
Change-Id: Ib57bc66f91686529beabc837bfd90a8c1ba7e42e