Support user action pseudo classes in non-subject :has()
To support user action pseudo classes in non-subject :has(), the
dynamic restyle flags for :has() invalidation were moved from the
computed style extra fields to ElementSuperRareData.

This adds a byte to the ElementSuperRareData. To fill up the added
byte, the flags in the ElementRareData were also moved.

The computed style extra field flag for pseudo state change was also
moved to get the same advantage of filtering with the flag for non-
subject :has()

Bug: 669058
Change-Id: I00f892c0c494c31cce197345669b403bf63cf6b2