More incremental style updates.
If we have an existing computed style on a given element, and the only
changes have been mutations of independent properties on the element's
inline style we can often reuse the old computed style and just reapply
the element's inline style on top of it. This allows us to skip
collecting elements and computing the full cascade, which can be a
significant win when animating elements via inline style from
JavaScript.

Speeds up the MotionMark “multiply” test by 20–27%, depending on
platform.

Change-Id: I78f1f30440cdb574dba9012b28fe27b1f60ef89f